assume a root level sidebar json for all pages, until later refactor is required
this simplifies the code and removes an unused path of code

Also, remove sidebar layout validation as this means that you cannot add a layout/sidebar
with updating the validator. As the validator was used in the filter function it also meant
that it determined whether or not a sidebar.json file was generated for the root index page.