feat: ensure snapshots are traced when deploying to vercel
When deploying a mosaic site to vercel, snapshot files are ignored resulting in a 404 when a page is accessed from the site.

Vercel uses [output file tracing](https://nextjs.org/docs/advanced-features/output-file-tracing) to determine what files need to be deployed. Due to the dynamic nature of mosaic sites, files that are required, the snapshots, are not traced correctly.

 When the deployment configuration below is added to a mosaic config file, mosaic will ensure that all snapshot files are included in the vercel output trace so that they are deployed.

 ```
  deployment: { mode: 'snapshot-file', platform: 'vercel' }
 ```