fix: `SharedConfigPlugin` uses closest shared config
### Before
If a page did not specify any shared config metadata then it would simply create a ref to the parent directory shared-config.json file.

However, this assumes that the parent directory *has* a shared config file.

### After
All generated shared config files are identified and then the closest 1 to the page is used.